create_readme {startR}	R Documentation

Creat README

Description

Creates a README.md file containing the structure of the repository and a footer

Usage

create_readme(repo = "Repository",
  footer = "<a href=\"https://orcid.org/0000-0003-1245-589X\" target=\"orcid.widget\" rel=\"noopener noreferrer\" style=\"vertical-align:top;\"><img src=\"https://orcid.org/sites/default/files/images/orcid_16x16.png\" style=\"width:1em;margin-right:.5em;\" alt=\"ORCID iD icon\">orcid.org/0000-0003-1245-589X</a>")

Arguments

repo
(character) The repository name, or any other title, defaults to 'Repository'

footer
(character) Text to be included in the footer. Defaults to my orcid

Examples

## Not run: 
create_readme() # Creates a readme file

## End(Not run)

[Package startR version 0.1.0 Index]