get_table {startR}	R Documentation

Get tables from GoogleBigQuery

Description

This function downloads tables from GoogleBigQuery and brings them into memory as a normal r data.frame. The first time it is ran, you will have to authenticate into your GBQ account.

Usage

get_table(project = "ucsb-gfw", dataset = NULL, table = NULL,
  allowLargeResults = TRUE, show_tables = TRUE)

Arguments

project
Name of the project, defaults to ucsb-gfw

dataset
Name of the dataset where the table lives

table
Name of the table to download

allowLargeResults
Allow large results?

show_tables
Show all tables within the dataset before downloading?

Value

An object of clas tbl.df

Examples

my_data <- get_table(dataset = "foreign_fishing_ren",
                     table = "annual_foreign_fishing")

[Package startR version 0.1.0 Index]