render_doc {startR}	R Documentation

Render Document

Description

Calls rmarkdown::render() to produce the output of a given rmd file

Usage

render_doc(file, rmd_dir = "rmd", output_dir = "docs", ...)

Arguments

file
A character vector specifying the file name

rmd_dir
The directory where the file lives, specified as a character vector

output_dir
The output directory, where the knitted versions will be placed

Examples

I have a file called \code{"manuscript.rmd"} that lives in my \code{"repo/rmd/"} directory.
## Not run: 
render_doc(file = "manuscript.rmd")

## End(Not run)

[Package startR version 0.1.0 Index]