sfc_as_cols {startR}	R Documentation

sfc_as_cols

Description

Take the geometry column of an sf point object and turn it into columns of the data.frame

Usage

sfc_as_cols(x, names = c("lon", "lat"))

Arguments

x
An object of class df sf

names

Value

a data.frame

[Package startR version 0.1.0 Index]