usual_suspects {startR}	R Documentation

Package loading

Description

Loads the usual packages from my every-day workflow

Usage

usual_suspects(pkgs = NULL)

Arguments

pkgs
A character vector containing the packages that should be loaded. The fedault ones include janitor, here, sf, and tidyverse

[Package startR version 0.1.0 Index]